+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 
 ### Fixed
 - Fixed locale change when accessing user edit page
+- Fixed Project delete
 - Fixed `story-description` and `.note` by adding to sortable column canceled elements/nodes
+
 
 ### Changed
 - pusherSockets.js to allow empty pusher env vars